spec: add spec for conditional format
Add basic spec for conditional formats. Also refactor package structure
a bit.
