+++ sheetwhat/checks/has_equal_conditional_formats.py
